refactored code for confirmation page
- moved confirmation page logic to custom hook
- moved styling into its own file
