+++ swimster-ui/src/components/ConfirmReservation/style.js
@@ -1,0 +1,45 @@
+export const style = {
+    confirmReservation: 'grid grid-cols-3 gap-x-5 mx-48 mt-20 font-normal text-md',
+    reservationDetailsContainer: 'col-span-2 space-y-10 p-7 rounded-xl shadow-md',
+    reservationDetailsHeader: 'font-medium text-2xl',
+    confirmReservationListingContainer: 'flex space-x-4',
+    reservationDateContainer: '',
+    confirmReservationListingImage: 'w-48 h-32 rounded-lg',
+    confirmReservationListingContent: 'flex flex-col space-y-5',
+    confirmReservationListingPrimary: '',
+    confirmReservationListingPoolType: 'font-medium text-gray-400 text-sm',
+    confirmReservationListingTitle: 'font-semibold text-lg',
+    confirmReservationListingAddress: 'font-semibold text-gray-400',
+    confirmReservationListingSecondary: '',
+    confirmReservationListingHostDetails: 'font-medium',
+    reservationDateHeader: 'text-lg font-semibold',
+    reservationDate: '',
+    reservationDurationContainer: '',
+    reservationDurationHeader: 'text-lg font-semibold',
+    reservationDuration: '',
+    reservationGuestsContainer: '',
+    reservationGuestsHeader: 'text-lg font-semibold',
+    reservationGuests: '',
+    paymentDetailsContainer: 'col-span-1 flex flex-col space-y-8 p-7 rounded-xl shadow-md',
+    paymentDetailsHeader: 'font-medium text-2xl',
+    poolCostContainer: '',
+    poolCostHeader: 'font-medium',
+    poolCost: '',
+    confirmReservationDurationContainer: '',
+    confirmReservationDurationHeader: 'font-medium',
+    confirmReservationDuration: '',
+    transactionDetailsBreak: 'h-0.5 bg-gray-200',
+    reservationSubtotalContainer: 'flex justify-between',
+    reservationSubtotalHeader: 'font-medium text-lg',
+    reservationSubtotal: 'text-lg',
+    serviceFeesContainer: 'flex justify-between',
+    serviceFeesHeader: 'font-medium',
+    serviceFees: '',
+    taxesContainer: 'flex justify-between',
+    taxesHeader: 'font-medium',
+    taxes: '',
+    totalContainer: 'flex justify-between',
+    totalHeader: 'font-medium text-xl',
+    total: 'text-lg',
+    confirmReservationButton: 'bg-indigo-400 rounded-lg py-3 px-2 text-gray-50',
+}
